+++ back/functions/requirements.txt
@@ -1,3 +1,5 @@
+annotated-types==0.7.0
+anyio==4.8.0
 blinker==1.9.0
 CacheControl==0.14.2
 cachetools==5.5.1
@@ -9,6 +11,8 @@ cloudevents==1.9.0
 colorama==0.4.6
 cryptography==44.0.1
 deprecation==2.1.0
+distro==1.9.0
+et_xmlfile==2.0.0
 firebase-admin==6.6.0
 firebase-functions==0.1.2
 Flask==3.1.0
@@ -27,24 +31,41 @@ google-resumable-media==2.7.2
 googleapis-common-protos==1.67.0
 grpcio==1.70.0
 grpcio-status==1.70.0
+h11==0.14.0
+httpcore==1.0.7
 httplib2==0.22.0
+httpx==0.28.1
 idna==3.10
 itsdangerous==2.2.0
 Jinja2==3.1.5
+jiter==0.8.2
 MarkupSafe==3.0.2
 msgpack==1.1.0
+numpy==2.2.3
+openai==1.63.2
+openpyxl==3.1.5
 packaging==24.2
+pandas==2.2.3
 proto-plus==1.26.0
 protobuf==5.29.3
 pyasn1==0.6.1
 pyasn1_modules==0.4.1
 pycparser==2.22
+pydantic==2.10.6
+pydantic_core==2.27.2
 PyJWT==2.10.1
 pyparsing==3.2.1
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+pytz==2025.1
 PyYAML==6.0.2
 requests==2.32.3
 rsa==4.9
+six==1.17.0
+sniffio==1.3.1
+tqdm==4.67.1
 typing_extensions==4.12.2
+tzdata==2025.1
 uritemplate==4.1.1
 urllib3==2.3.0
 watchdog==6.0.0
